pyproject.toml
[project]
name = "twipsybot"
version = "0.1.0"
description = "AI-powered Misskey bot with OpenAI-compatible LLMs."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

dependencies = [
    "aiohttp==3.13.3",
    "aiosqlite==0.22.1",
    "anyio==4.12.0",
    "apscheduler==3.11.2",
    "beautifulsoup4==4.14.3",
    "cachetools==6.2.4",
    "click==8.3.1",
    "feedparser==6.0.11",
    "humanfriendly==10.0",
    "loguru==0.7.3",
    "openai==2.14.0",
    "pydantic==2.12.5",
    "psutil==7.2.1",
    "pytimeparse2==1.7.1",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.1",
    "tenacity==9.1.2",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.408",
]

[project.scripts]
twipsybot = "twipsybot.app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["data", "logs"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["twipsybot"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["twipsybot", "plugins"]
exclude = ["data", "logs"]
reportUnsupportedDunderAll = "none"
reportAttributeAccessIssue = "error"
executionEnvironments = [
    { root = "plugins", reportAttributeAccessIssue = "none" },
    { root = "twipsybot/clients/misskey", reportAttributeAccessIssue = "none" },
]
